requirements.txt
discord.py~=2.6.2
python-dotenv~=1.1.1
motor~=3.7.1
requests~=2.32.5
tabulate~=0.9.0